+++ tests/fixtures/compiler/includes/CompetingInclude.pas
@@ -1,0 +1,11 @@
+unit CompetingInclude;
+
+{$mode objfpc}{$H+}
+
+interface
+
+{$I choice.inc}
+
+implementation
+
+end.