+++ tests/fixtures/compiler/includes/LocalIncludeAPI.pas
@@ -1,0 +1,11 @@
+unit LocalIncludeAPI;
+
+{$mode objfpc}{$H+}
+
+interface
+
+{$I local.inc}
+
+implementation
+
+end.